Drop sale_converter file.
