+++ saleor/graphql/discount/mutations/sale/sale_update.py
@@ -8,7 +8,6 @@ from django.core.exceptions import ValidationError
 from .....core.tracing import traced_atomic_transaction
 from .....discount import models
 from .....discount.error_codes import DiscountErrorCode
-from .....discount.tests.sale_converter import create_catalogue_predicate
 from .....discount.utils import CATALOGUE_FIELDS
 from .....permission.enums import DiscountPermissions
 from .....product.tasks import update_products_discounted_prices_for_promotion_task
@@ -29,6 +28,7 @@ from ....plugins.dataloaders import get_plugin_manager_promise
 from ...types import Sale
 from ...utils import (
     convert_migrated_sale_predicate_to_catalogue_info,
+    create_catalogue_predicate,
     get_products_for_rule,
 )
 from .sale_create import SaleInput
